# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type SignupBodySelectedCharacter =
  (typeof SignupBodySelectedCharacter)[keyof typeof SignupBodySelectedCharacter];

export const SignupBodySelectedCharacter = {
  "infinity-mentor": "infinity-mentor",
  "dark-king": "dark-king",
  "energy-hero": "energy-hero",
} as const;

export interface SignupBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  selectedCharacter?: SignupBodySelectedCharacter;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserSelectedCharacter =
  (typeof UserSelectedCharacter)[keyof typeof UserSelectedCharacter];

export const UserSelectedCharacter = {
  "infinity-mentor": "infinity-mentor",
  "dark-king": "dark-king",
  "energy-hero": "energy-hero",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  selectedCharacter: UserSelectedCharacter;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UpdateCharacterBodySelectedCharacter =
  (typeof UpdateCharacterBodySelectedCharacter)[keyof typeof UpdateCharacterBodySelectedCharacter];

export const UpdateCharacterBodySelectedCharacter = {
  "infinity-mentor": "infinity-mentor",
  "dark-king": "dark-king",
  "energy-hero": "energy-hero",
} as const;

export interface UpdateCharacterBody {
  selectedCharacter: UpdateCharacterBodySelectedCharacter;
}

export interface Habit {
  id: number;
  userId: number;
  title: string;
  description?: string;
  completedDates: string[];
  currentStreak: number;
  longestStreak: number;
  createdAt: string;
}

export interface CreateHabitBody {
  title: string;
  description?: string;
}

export interface UpdateHabitBody {
  title?: string;
  description?: string;
}

export interface CompleteHabitBody {
  date: string;
}

export interface WeeklyDataPoint {
  date: string;
  completed: number;
  total: number;
}

export interface DashboardSummary {
  totalHabits: number;
  completedToday: number;
  completionRate: number;
  longestStreak: number;
  totalCompletions: number;
  weeklyData: WeeklyDataPoint[];
}

export type InsightType = (typeof InsightType)[keyof typeof InsightType];

export const InsightType = {
  timing: "timing",
  consistency: "consistency",
  streak: "streak",
  warning: "warning",
  encouragement: "encouragement",
} as const;

export interface Insight {
  id: string;
  type: InsightType;
  message: string;
  habitId?: number;
  habitTitle?: string;
}

export interface HabitStreak {
  habitId: number;
  habitTitle: string;
  currentStreak: number;
  longestStreak: number;
  lastCompletedDate?: string;
}

export type CompanionMessageCharacter =
  (typeof CompanionMessageCharacter)[keyof typeof CompanionMessageCharacter];

export const CompanionMessageCharacter = {
  "infinity-mentor": "infinity-mentor",
  "dark-king": "dark-king",
  "energy-hero": "energy-hero",
} as const;

export type CompanionMessageMood =
  (typeof CompanionMessageMood)[keyof typeof CompanionMessageMood];

export const CompanionMessageMood = {
  encouraging: "encouraging",
  challenging: "challenging",
  disappointed: "disappointed",
  celebrating: "celebrating",
  neutral: "neutral",
} as const;

export interface CompanionMessage {
  character: CompanionMessageCharacter;
  message: string;
  mood: CompanionMessageMood;
  completionRate: number;
}