# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const signupBodyPasswordMin = 6;

export const SignupBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  password: zod.string().min(signupBodyPasswordMin),
  selectedCharacter: zod
    .enum(["infinity-mentor", "dark-king", "energy-hero"])
    .optional(),
});

/**
 * @summary Login user
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    selectedCharacter: zod.enum([
      "infinity-mentor",
      "dark-king",
      "energy-hero",
    ]),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  selectedCharacter: zod.enum(["infinity-mentor", "dark-king", "energy-hero"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update selected character
 */
export const UpdateCharacterBody = zod.object({
  selectedCharacter: zod.enum(["infinity-mentor", "dark-king", "energy-hero"]),
});

export const UpdateCharacterResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  selectedCharacter: zod.enum(["infinity-mentor", "dark-king", "energy-hero"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get all habits for current user
 */
export const GetHabitsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  completedDates: zod.array(zod.coerce.date()),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  createdAt: zod.coerce.date(),
});
export const GetHabitsResponse = zod.array(GetHabitsResponseItem);

/**
 * @summary Create a new habit
 */
export const CreateHabitBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
});

/**
 * @summary Update a habit
 */
export const UpdateHabitParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateHabitBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().optional(),
});

export const UpdateHabitResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  completedDates: zod.array(zod.coerce.date()),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a habit
 */
export const DeleteHabitParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Mark habit as complete for a date
 */
export const CompleteHabitParams = zod.object({
  id: zod.coerce.string(),
});

export const CompleteHabitBody = zod.object({
  date: zod.coerce.date(),
});

export const CompleteHabitResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  completedDates: zod.array(zod.coerce.date()),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Mark habit as incomplete for a date
 */
export const UncompleteHabitParams = zod.object({
  id: zod.coerce.string(),
});

export const UncompleteHabitBody = zod.object({
  date: zod.coerce.date(),
});

export const UncompleteHabitResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  completedDates: zod.array(zod.coerce.date()),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalHabits: zod.number(),
  completedToday: zod.number(),
  completionRate: zod.number(),
  longestStreak: zod.number(),
  totalCompletions: zod.number(),
  weeklyData: zod.array(
    zod.object({
      date: zod.string(),
      completed: zod.number(),
      total: zod.number(),
    }),
  ),
});

/**
 * @summary Get AI-powered habit insights
 */
export const GetInsightsResponseItem = zod.object({
  id: zod.string(),
  type: zod.enum([
    "timing",
    "consistency",
    "streak",
    "warning",
    "encouragement",
  ]),
  message: zod.string(),
  habitId: zod.number().optional(),
  habitTitle: zod.string().optional(),
});
export const GetInsightsResponse = zod.array(GetInsightsResponseItem);

/**
 * @summary Get streak data for all habits
 */
export const GetStreaksResponseItem = zod.object({
  habitId: zod.number(),
  habitTitle: zod.string(),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  lastCompletedDate: zod.string().optional(),
});
export const GetStreaksResponse = zod.array(GetStreaksResponseItem);

/**
 * @summary Get character companion message
 */
export const GetCompanionMessageResponse = zod.object({
  character: zod.enum(["infinity-mentor", "dark-king", "energy-hero"]),
  message: zod.string(),
  mood: zod.enum([
    "encouraging",
    "challenging",
    "disappointed",
    "celebrating",
    "neutral",
  ]),
  completionRate: zod.number(),
});