# lib/api-zod/src/generated/types/companionMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanionMessageCharacter } from "./companionMessageCharacter";
import type { CompanionMessageMood } from "./companionMessageMood";

export interface CompanionMessage {
  character: CompanionMessageCharacter;
  message: string;
  mood: CompanionMessageMood;
  completionRate: number;
}