# lib/api-zod/src/generated/types/createHabitBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateHabitBody {
  title: string;
  description?: string;
}