# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */
import type { WeeklyDataPoint } from "./weeklyDataPoint";

export interface DashboardSummary {
  totalHabits: number;
  completedToday: number;
  completionRate: number;
  longestStreak: number;
  totalCompletions: number;
  weeklyData: WeeklyDataPoint[];
}