# lib/api-zod/src/generated/types/habit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */

export interface Habit {
  id: number;
  userId: number;
  title: string;
  description?: string;
  completedDates: Date[];
  currentStreak: number;
  longestStreak: number;
  createdAt: Date;
}