# lib/api-zod/src/generated/types/habitStreak.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */

export interface HabitStreak {
  habitId: number;
  habitTitle: string;
  currentStreak: number;
  longestStreak: number;
  lastCompletedDate?: string;
}