# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */
import type { SignupBodySelectedCharacter } from "./signupBodySelectedCharacter";

export interface SignupBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  selectedCharacter?: SignupBodySelectedCharacter;
}