# lib/api-zod/src/generated/types/updateCharacterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCharacterBodySelectedCharacter } from "./updateCharacterBodySelectedCharacter";

export interface UpdateCharacterBody {
  selectedCharacter: UpdateCharacterBodySelectedCharacter;
}