# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowHabits API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSelectedCharacter } from "./userSelectedCharacter";

export interface User {
  id: number;
  name: string;
  email: string;
  selectedCharacter: UserSelectedCharacter;
  createdAt: Date;
}